+++ CHANGELOG.md
@@ -191,3 +191,4 @@
 * `Binary` trees can be empty
 * Define `left_sub_tree` and `right_sub_tree` lenses of `Binary` tree
 * Fix `insert` function: `Binary` tree should not contain duplicate values
+* Remove `Cartesian` type class